WIP: remove alias command and create get-aliases command Done this because the hacked alias command I previously committed was causing linting errors because of unused vars, and this was the simpler solution
